include all common commands in v2
